Address review comments: move Puffin DV interop test to test_deletes.py
- Move test_read_spark_written_puffin_dv from test_puffin_spark_interop.py
  to test_deletes.py; delete the old file
- Replace run_spark_commands helper calls with direct spark.sql() calls
- Tighten entries assertion from > 0 to == 1
